architecture: каркас DDD/Hexagonal — перенос AppServiceProvider и базового Controller в Shared; перенос User в модуль; правки фабрики/сидера и Makefile
