+++ app/Http/Controllers/Controller.php
@@ -2,7 +2,7 @@
 
 namespace App\Http\Controllers;
 
-abstract class Controller
+abstract class Controller extends \App\Shared\UI\Http\Controller
 {
     //
 }
