+++ app/Shared/Infrastructure/Providers/AppServiceProvider.php
@@ -1,6 +1,8 @@
 <?php
 
-namespace App\Providers;
+declare(strict_types=1);
+
+namespace App\Shared\Infrastructure\Providers;
 
 use Illuminate\Support\ServiceProvider;
