+++ config/auth.php
@@ -62,7 +62,7 @@ return [
     'providers' => [
         'users' => [
             'driver' => 'eloquent',
-            'model' => env('AUTH_MODEL', App\Models\User::class),
+            'model' => env('AUTH_MODEL', \App\Users\Infrastructure\Database\Eloquent\Models\User::class),
         ],
 
         // 'users' => [
